+++ flowjax/nn/__init__.py
@@ -1,4 +1,4 @@
-from .block_autoregressive_linear import BlockAutoregressiveLinear
+from .block_autoregressive import BlockAutoregressiveLinear
 from .masked_autoregressive import AutoregressiveMLP, MaskedLinear
 
 __all__ = [
